Move sha1 to mod hash, use anyhow
