+++ app/src/main/rust/Cargo.toml
@@ -18,7 +18,8 @@ once_cell = { default-features = false, version = "1.19" }
 serde_cbor = "0.11"
 serde = { version = "1.0", features = ["derive"] }
 sha1 = { version = "0.10", features = ["asm"] }
-base16ct = { version = "0.2" }
+base16ct = { version = "0.2", features = ["std"] }
+anyhow = { version = "1.0" }
 
 [profile.release]
 panic = "abort"
