+++ app/src/main/rust/src/hash/mod.rs
@@ -1,0 +1,1 @@
+mod sha1;
